Changed so a Course is associated with a proposal instead of a project
