+++ app/Models/Course.php
@@ -6,5 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Course extends Model
 {
-    
+	public function proposals()
+	{
+		return $this->hasMany(Project::class);
+	}
 }
